requirements.txt
apache-airflow==2.5.3
Faker==18.4.0
apache-airflow-providers-postgres==5.4.0